serial speed; bug fixes; general improvements

+++ oquam/Oquam/encoder.h
@@ -6,6 +6,6 @@ void init_encoders();
 void handle_x_encoder_interrupt();
 void handle_y_encoder_interrupt();
 
-extern volatile int16_t encoder_position[3];
+extern volatile int32_t encoder_position[3];
 
 #endif // _OQUAM_ENCODER_H_
